dded AJAXiness to upgrading settlement to city.
Previously, had problem with links to action partials. If, for example,
a specific settlement was not available to upgrade to a city, the url
generator would raise an exception. So we created
link_to_action_partial_if to not execute the code that generates the
url. But with AJAX, the link had to be available at all times in case an
action would cause that link to be necessary. Also, the link could not
be specific to a settlement, it had to be generic. In the process of
changing this in this commit.